fix: remove check volume backed filesystem warning
When the storage provisioner reflects back to the controller the
filesystem provisioned info, it is not possible for it to pick
which volume backs it. This is chosen ahead of time and returned
in the FilesystemParams facade method. The reason for the this
volume value being sent to the controller from the storage
provisioner worker is from poor re-use of params structures.