test: take http context tests out of _test package
This commit removes all of the httpcontext package tests out of a sub
test package. We are moving away from sub test packages and this is
required to properly test new functionality being added to http context.

It is also one of the reasons why we have to export functions in this
package that would otherwise not have to be exported.