feat: introduce a new handler for signaling controller model
This commit introduces a new handler to the apiserver for signaling on a
requests context if the request is for the controller model. This has
been introduced to help provide stronger authz into the apiserver.

While we wish for a day where the controller model can hopefully stop
being special cased we must maintain a strong authz posture now with
regards to the controller model.

This work was done to support controller upgrades in Juju 4.0. In Juju
4.0 we want to run a different set of permissions checks if the model
being worked on is the controller model.