refactor: remove the controller tools upload handler
We had a special tools upload handler for the controller that would
allow tools upload into the controllers object store. In reality the
juju client never calls this handler.

Tools upload is always performed to a model. To help reduce our API foot
print and the potential security issues this commit removes the handler
as it is no longer needed.