refactor: model agent binary store for tools request
Refactors the model agent binary store helper for tools request to
consider of the tools request is for the controller's model. If this is
the case the caller will be given the controller's agent binary store
instead of the model's.

This is done because when tools uploading is done we need to be able to
distrubute the controller's tools out to all models under the
controller.

This can only happen when the controller's objects store and not the
model has a cache of these tools.

In an ideal world this logic would be pushed down further into the
domain services.