fix: return not provisioned error if block device does not have a name
The storageprovisioner worker uses the device name (e.g. sda) to generate
a device path (e.g. /dev/sda). If it is not currently set for a block
device, that means it is a partial block device entry. By returning
a NotProvisioned error, the storageprovisioner worker will retry later.

The block device will eventually have a device name if the diskmanager
worker sends matching block device information.

Without this fix, the storageprovisioner will try to create a partition
table, and fail.

Fixes #21073.