fix(tests/secrets_iaas): fix race in secret suite CI tests
This patch removes the `--config token=foo` parameter on deployment of `juju-qa-dummy-source`
in `check_secrets` to replace it by an explicit call after deployment. This ensures that the start
event won't race with the config-changed event. Having the start event after the config-changed
event causes the application to stay in the wrong status which break the CI assertion.

I also replace `juju-qa-dummy-source`by a simple `juju-qa-test` where the relation isn't required.