test: gives the user super user access in tools tests
We have an integration test for macaroons that broke because it is using
the tools endpoints on the controller model and did not have super user
access permissions.

I have given the correct permissions in the tests because the test if
focused on asserting behaviour of tools and authorization. The
authorization logic we have has already been covered well with tests.

In an ideal world we would get rid of this test suite has it has very
low value.