fix: stops the provider tracker worker form crashing when setting model config
When changing model config the provider tracker will crash and come back
up. This is because the model config passed to the environ does not
contain an agent version. When getting parsed by model config the
validation sees this as removing the agent version and throws an error.

Juju 4.0 does not allow setting or changing agent version via model
config anymore. We can safely remove this check as it adds no value.
Callers getting agent version from model config should be changed.