refactor: change the type used for storage directive max count
When the storage directive struct was created in the application domain
it was setup to use a unit32 value with the safety implied that a
negative value could be obtained.

In reality we expect the charm to supply a value of -1 when no maximum
storage count exists.

This change changes the type to an in on StorageDirective and defines a
new const to be used by business logic for testing this case.