fix: avoid panic on firewaller when relation removed
This patch fixes a panic on the firewaller, which is fired by the names library
when receiving an empty relation key.

The fix is to actually avoid getting the tag altogether by replacing the
map index with a relation UUID, which is consistent to the changes we
receive from the watcher.