fix(ci): simplify secret ID handling for CI tests
There are few changes on how the status log is printed and how the juju_report_engine is
layered.

This patch fixes all formatting issues and tries to rely a bit less on formatting.

Also, it removes the need to install yq onto the deployed machine by moving the report processing
onto the CI job, which already has yq.