feat: re-implement remote model UUID for CMR
This patch fixes (by re-implementing) relation-model-get hook.

The problem was that we had removed a check for CMRs, this patch
implements getting the remote model UUID for a given relation UUID. If
the given relation is not CMR then the API will simply not update the
model UUID for the remote application. This is the same behavior as 3.6.