refactor: return not yet supported errors for storage facade
There are still several funcs in the client storage facade that have not
been implemented for a Juju 4.0 release. To help with the 4.0 release
this commit stubs out these funcs returning a not yet available params
code.

I chose this code because of the error message the client produces.
Specifically it lets the user know that it isn't available in the
version of the controller.

If we return a not implemented error it produces a more complex error
message that makes the user think they don't have the right client
version.