feat: delete model db
Use the new dqlite delete the database pragma. This unlocks the ability
to delete a model correctly without having to iterate through each
table, trigger, view et al. This should make juju a lot more correct
and performant in an existing production setup.

Interestingly, we have to keep the old deletedb workaround for tests
so that we can actually delete the database. This is only locally and
we could just skip them if run locally. For now, I've kept them in
for backwards compatibility.