feat: introduce a third provisioning scope into the Juju model
This commit introduces a third provisioning scope into the Juju model
for talking about storage that is externally provisioned to the
controller. That is the provider and provisioner of the storage does not
run under Juju's normal control loop.

This is a draft idea in support of some storage work being done by Harry
for k8s. It may not land but aids to seed discussion if we want to
include this in a 4.0 release at the start.