fix: remove LXD profile functionality from uniter
Since Juju 4.0 does not support LXD profiles, we need to prevent the
situation where the unit agent breaks if one is declared in the charm's
metadata.

Here we remove the main machinery from the uniter, so it does not watch
or attempt to verify and apply profiles. There are still API client
capabilities for this functionality untouched. These will be removed
later.