feat: add one to the tracker connection pools
The limiting of the connection pool to 2 was used as a defacto semaphore
to prevent a stampeding herd as we approached the limit of Dqlite and
began retrying all transactions.

Experiments show that 3 is significantly faster than 2, with no evident
ill-effects.

We will persue performance further, but at this time we take the easy
win.