fix: application's creation config type
validateCreateApplicationArgs was validating and mutating the
arg.ApplicationConfig, however we were passing arg as a value and not
returning the mutated struct. This caused a bug because
makeApplicationArg relies on the arg.ApplicationConfig to have the right
go type to create the Config struct. Now the validate function returns
the mutated arg.