fix: application worker for caas firewaller safe shutdown
This commit fixes the application worker for caas firewaller to
correctly shut down the worker and watcher dependencies when an
application not found error is received.

When the worker receives this error it means the application the worker
was watching has been removed from the model. In this case it is not an
error and the worker just needs to safely shut down and take any
watchers with it.

This fixes a bug where the worker was swapped over to dqlite and the
error was not updated to handle this case.