test: adds regression test for worker shut down
Adds missing tests to the application worker for the caas firewaller to
ensure that when the application is removed from the model the worker
shuts down cleanly with no and removes child workers.

I took the time here to clean the test suite up a bit and remove gomock
in order usage.