fix: stop throwing away not found error from broker
The caas firewaller when calling off to the Kubernetes broker was
checking if the error was not found and then throwing away the error and
continuing on with normal execution of the worker.

This shouldn't be happening, if the broker is telling us the application
does not exist in the environ then this is an error that should be
surfaced.