refactor: correctly order logic of caas firewaller worker
This commit refactors the logic of the main firewall worker for the caas
firewaller. It now correctly handles application not found errors and
starts and stops child workers accordingly.

It also checks the application life first before any other checks. This
is a better approach as we can avoid more database transactions if an
early escape path exists.

I have also added better logging that works with the child logger and
composed the problem into smaller units.