refactor: remove transiant dependecies from firewall worker
The firewall worker has to start child application workers and there for
carries the child worker dependency with it. This doesn't need to happen
as this is the reason we have manifolds.

The worker just requires a closure to call which has captured the
dependencies avoid the need for the main firewall worker ever having to
care.