fix: logic error in starting app worker
Fixes a logic error with ensure app worker for caas firewaller. I was
only returning out of the func when it should have always been doing
this.

While I here I renamed the config to FirewallerConfig.

Also made the check for charm v2 only happen if the application uuid is
not already knwon to the worker.