feat: schedule controller model removal job
We need to postpone the controller model removal job until all hosted
models have been removed. We can simply check what model UUIDs are in
the controller model as a signal for this. If the model uuids are empty
(after the removal of the current model UUID), then we can schedule
a model removal job. **note** This is the only time we're allowed to
schedule a job from another job. We didn't want this to happen
originally as it can cause non-deterministic behaviour (similar to
how 3.6 works). As this is the _only_ place we're allowing this, this
should be ok. The alternative is to duplicate all removal jobs for
the controller model, which seems wasteful.