fix: handle container and parent from args to AddMachines
Bundles supply machine arguments for containers by populating
ContainerType and ParentId in AddMachineParams.

Prior to Juju 4, if these were populated, we would call a specific state
method to add machine-in-machine. We would also set them from incoming
placement directives if appropriate.

We no longer do this. We call one method from the domain to add
machines, and that method handles all forms of placement. So we need to
invert the prior logic and populate placement from these other fields.