fix(ci): add CA certificate to secret backend setup
This patch updates `tests/suites/secrets_iaas/vault.sh` to include the CA certificate when
configuring the Vault backend using `juju add-secret-backend`.

The CA certificate is read from `$VAULT_CAPATH` and added to the `ca-cert` parameter.

It is required because without this config param, adding the backend would fail with a TLS error,
due to the self-signed certificate.