fix: do not retry transaction rollbacks
The first call to Commit or Rollback sets the transaction as done as the
very first step. There is no point retrying a failed rollback - it will
never succeed.

In addition, we do not swallow ErrTxDone errors on commit, as we should
retry such transactions if we don't know if they performed the desired
state mutations.

We also trace rollbacks in the same way as commits.