fix(secretbackend): support complex config values
This patch adds support for encoding and decoding complex configuration values in
 `domain/secretbackend` ensuring enhanced flexibility and robustness.

Before this patch, an array was encoded as a string with `fmt.Sprintf("%+v")` and thus, not
symmetrically decoded. This causes the TLS certificate verification fails for k8s secret provider,
because the logic wasn't able to properly decode the content of `ca-certs` configuration key, which
contains a slice of string, each string as a unique certificate.

After this patch, we use json.Marshal/json.Unmarshal to insert/select values from the secret backend
configuration. This fixes the issue with TLS verification, which was triggered in CI suite
`secrets_iaas/test_secrets_k8s`

If a configuration is malformed, cannot be marshalled or unmarshalled, a log will be produced, and
the key will be ignored.