feat(actions): add action to disallow DDL changes
To make changes to the domain DDL between Juju patch versions, we cannot
make changes directly as this will cause controller upgrades to break.
Instead, we use .PATCH.sql files

Add a github action that assures not changes are made to the DDL with
the exception of .PATCH.sql files

Skip this check on the main branch, since we use that branch to develop
the next, unreleased minor version, where changes to the DDL are
allowed.