feat: introduce state machine for provisioner task states
This patch is the first out of many that aims to improve the readability
and reduce races in the provisioner task.

This commit introduces a new FSM for modelling the machines' states in
the context of the provisioner task.

The main goal here is to get rid of the legacy "defer stop" path.
Although this is not done in this commit, the introduction of a
CancellingStart state will allow us to rewrite that logic without locks
and mutable variables.