fix(secretbackend): immutability for built-in backends
This patch fixes the database triggers which ensure "internal" and "kubernetes" secret
backends are immutable.

Before this patch, these triggers used the backend_type_id to ensure immutability, asserting that
backend with types "controller" and "kubernetes" cannot be removed. This caused an issue when
someone adds a new backend on k8s as done in test suites `secrets_iaas test_secrets_k8s`

After this patch, the trigger uses the names (`internal` and `kubernetes`) which fixes the issue.

Note: this patch is meant to be merged in the released 4.0 branch, for main it should be done
in `domain/schema/controller.go` directly.