feat: split into smaller funcs and introduce a Migrator interface
We were directly calling a concrete implementation of migration.ImportModel which makes it hard to test the storage unique ID migration changes (specific for CAAS). The Migrator interface will help us mock that implementation in unit tests. While the existing migration.ImportModel is still there for IAAS, at least this alleviates the testing for CAAS. Refactoring the migration.ImportModel for IAAS is for a future effort.