fix: use sqlair Preparer for space constraints
This patch addresses an issue that prevents deploying an app using space
constraints. The reason for this being that we have the same db struct
name in two different packages, which creates a conflict on the prepared
statements cache. The fix is therefore (as it was already done before)
to use sqlair.Prepare instead of the injected preparer.