fix(secretbackend): immutability for built-in backends
This patch introduces the concept of "origins" for secret backends to distinguish between built-in
 and user-defined entries.

Before this patch, these triggers used the backend_type_id to ensure immutability, asserting that
backend with types "controller" and "kubernetes" cannot be removed. This caused an issue when
someone adds a new backend on k8s as done in test suites secrets_iaas test_secrets_k8s

After this patch, a new `secret_backend_origin` table and association column have been introduced,
and the business logic relies on those origins to define if we can or cannot remove or update a
backend.

Triggers have been updated to also rely on origin instead of type, even if the business logic
doesn't rely on it anymore.