fix: consume details should return offer name
This patch addresses an issue where we cannot consume any offers because
the endpoint GetConsumeDetails wasn't returning the offer name.
The endpoint was refactored in https://github.com/juju/juju/pull/21161
and the offer name removed from the returning struct.

The fix is therefore to add the offer name taken from the offer url as
it was prior to #21161.