fix: makefile target for updating minikube
This commit fixes the Makefile target for updating the Juju operator
image in minikube. Specifically this was not working how we thought it
was as the minikube command requires the name of the image to be
imported and only works in the way we were using it when the minikube
cluster is backed by the same runtime that built the image.

I noticed that it wasn't working when I was using podman to build images
but was using containerd to run minikube.

This change now saves the image to a temp file and loads it into
minikube from the temp file which allows the run time and builder to be
separate.

I used a temp file not under the project dir just in case the process
fails and then at least the temp file will be cleaned up through a
normal system restart.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>