feat: update lxd dependency
There is LXD vulnerability GO-2025-4121 that has yet to be tagged and
released. This will come in the up coming week(s), so for now update
to the latest dependency with the fix, but without the tag in the
release. Then skip the vuln check for now, and we can circle back
around and remove that check at a later date.

 1. https://pkg.go.dev/vuln/GO-2025-4121