fix: task logs watcher test flakyness
The TestWatchTaskLogs test was failing in a flaky fashion due to the
harness test not reaching an idle state in the check/assertion phase.

This happened because the log insertion was done in separate
transactions (15 in this particular test), and we can only guarantee
that db changes within the same transaction will end up in the same
term.

The fix is to change the test so that we insert a batch of logs in a
single transaction.