feat: handle different controllers debug-log
Locating the current controller addresses was using the wrong method
for getting the current controller name. Essentially we bypassed the
base comment maybeInitModel, which then caused us to access the wrong
controller with the incorrect certificate.

The fix is to just ask the base command what it thinks is the current
controller and that will initialise the model and report back the
correct one. Then we can correctly log in to the right one without
the TLS certificate error.