test: fix flakey test in change stream test
There isn't a guarantee that the time after will always be the default
timeout, because the loop might have already processed the next event.
In this case, there already exists a different expect method for when
this happens, and we didn't use it in the test. This was a mistake.