feat: add Actions check to guard against invalid schema changes
Create a script called ddlgen that writes the ddls for all released
versions of juju for the current major.minor into the tree.

Use this to check that no invalid changes are made

We need this because, to support controller upgrades, the only changes we can make
to the DDL between patch releases are additive. Applying the DDL is idempotent
using stored hashes of the patches we've already applied. These hashes are
calculated before the DDL is parsed, meaning comments, whitespace, etc. also
cannot change.

This took a bit of refactoring of domain/schema meaning we now encode
directly which patch is applied to which version.