refactor: use worker runner for managing child workers
This commit moves the caas firewaller over to using the worker runner
implementation for managing child workers. Originally we were using a
map of workers to do this.

It makes more sense to use the tools we already have instead of a custom
implementation.

This work also required plumbing through wall clock for the runner and
also using the go sync package to sync the test conditions. This is
important as the runner is internal to the worker and there is not way
to perform sync operations with it.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>