refactor: remove worker runner from firewall composition
When adding the worker runner to the caas firewaller it was part of the
composition of the struct supplying the 'Report' method. This change
instead makes it a named member and implements a proxy Report func onto
the runner.

This was done as part of direct PR feedback and makes the code easier to
comprehend about what interface is being supplied.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>