feat: handle empty instance status
The provider will return an empty status code if the underlying machine
is stopped (for LXD, it's the same for other providers). Unfortunately,
we don't handle empty status types, so when the instance poller attempts
to set the status it will fail. The solution for now is to set it to
unknown. This will stop the instance poller bouncing. It doesn't
actually solve the issue of miss-reporting 100% of the time, because
it depends on which controller you're connected to when you execute
the command.

For now just stop the bouncing, we'll improve this in the future.