feat: new authentication interface for Juju
This commit makes a first pass draft at introducing a new authentication
interface within a Juju controller to establish trust between two
actors.

I started at this interface so as to best represent the desire of what
is required for authenticating. The interface purposelly does not talk
about the interface mechanisms for which two actors get authenticated
and leaves this as implementation detail for higher level components.