fix(cmd): Introduce dial timeout when reaching a controller in show-status-log
Before this patch, if a controller wasn't responding, the command will block for a long time,
instead of returning early with what it gets from the other controllers.
After this patch, if a controller is down and cannot be dialed (because it is down), the command
will simply return what it got from other controller.