feat: generate base export types from SQL schema
Introduces generation logic that parses the SQL schema files located
in domain/schema/model/sql/ and generates corresponding Go structs and
Sqlair statements.

A 'go:generate' directive is included to automate the generation
process, and a new 'generate-models' target has been added to the
Makefile to trigger this generation.

A new type, ModelDump, has been added to the end of the generated
model.go file. This type contains slices of all the generated types,
and will be used to dump the entire model.

This will constitute a format-agnostic export from which other
format-dependent exports can be generated, such as for migration,
back-up and terraform export.