feat: add new user name builder for parts
This commit adds a new user name builder that allows building up a
username based on parts such as the name and domain component. It allows
logic within the controller to handle the values seperately and not have
to concern itself with how they are composed together as one string.