feat: introduce new error condition to authenticator
This commit introduces a new error const to the internal/auth package
and a contract update to the Authenticator interface. This new error
constant allows for an authenticator to return an error that informs the
caller to stop any more authentication work flows from happening.

It may seem odd to perform the design in this way as you end up
programming by exception. However under this circumstance the
authenticator will always be in an error state. This helps the
authenticator to add context to the error and inform the caller that no
further action can be taken.