feat: add support for decorating a context with auth audit info
This commit adds a suggestion by harry to the AuthResult interface on
offer to consumers. The interface now allows the the holder of a
AuthResult to decorate a context with auth information for auditing
purposes.

This will be useful in tracing and logging when trying to understand
authentication flows within a Juju controller.

To support this work I beefed up some of the interface docs around
expected errors and the state of an AuthResult when a error is received.

In support of this several new functions have been added to for adding
audit information to contexts.