refactor: update jaas authenticator for new auth result interface
This commit updates the JAAS authenticator to meet the new AuthResult
interface. The new update decorates contexts with auth information about
the authentciated JAAS user.

On top of this it was needed to make the ensure operation for a
controller user only happen one per AuthResult. This allows all funcs of
the AuthResult interface to be used but not result in extra database
load for no reason.