refactor: update authresult for local user to add audit support
This commit adds to the localuser auth result a new interface
requirement to add audit information to a context. Because of this
AuthResult has been changed to a struct so it can carry more information
to support this need.

Tests have been added to assert the contracts on offer by AuthResult.