feat: introduce a new type for storage providers
This commit introduces a new type for the storage domain to represent a
storage provider type. There exists a value in the internal package but
we are looking to get off of it.

The value contained in the internal package also has strings attached
that we don't want to take on.

This was done as part of getting creating storage pools working again
via the facade.