refactor: create storage pool service args
This commit integrates the provider type value of the domain and removes
the type for storage pool attributes from the create storage pool func
on the storage domain.

This was done to bring the interface inline with out expectations of
service funcs. The custom type for attributes as not adding and value
and just masking the poor design of the attribute system.