feat: new storage pool name validator
This commit introduces a new storage pool name validator for the storage
domain. The new validator takes what Juju had already been using in the
internal package and modifies it to be both a bit more stricter and also
more liberal in some respects.

The old validator that we had introduced a copy and past bug from other
regex that ended up allowing question mark characters in storage pool
names. This was never the intented purpose. The old validator also had
no upper bounds on the maximum length of a storage pool name.

Because of this unbound property is possible to have a denial of service
path for storage pool names that could overwhelm parts of the system. To
be strict and secure we need to establish some sort of upper bound. It
was decided to use 128 runes.

The validator also now supports unicode letter classes instead of just
ASCII. This allows Juju to start moving to a more language agnostic
implementation supporting all types of users.

Because a change in validation of storage pool names has been introduced
it now makes us ask how will we support existing storage pools that get
model migrated into the controller. The simple answer is we will reject
them through validation and force the user to update the names in the
model before migrating.

This is the preferred approach as we don't need to carry transition
logic forever and gets the problem dealth with as soon as possible.