feat: reimplement storage pool creation
This commit re-instates storage pool creation via the domain service.
The new implementation is now safe for the values supplied and performs
the proper checks required.

I have taken the oppurtunity to also bolster up the contracts on offer
and make the func as reliable as possible. While here the service also
now both generates the storage pool uuid and returns it out of the func.

Test for storage pool creation was also done taking the oppurtuntiy to
clean up the existing testing that existed.