feat: create storage pool support in apiserver facade
This commit re-implements creating storage pools in the apiserver facade
again. I made the decision to split out storage pool related facade
methods into their own seperate file. This us allows us to think about
the client storage facade in terms of the individual components that
comprise storage.

I also cleaned up the testing in this package by making suites for the
individual parts of the facade and not relying on a contrived default
state from the base suite.