feat: auto upgrade model
When a model becomes active, before we remove the model importing
status, set the model target version. This should ensure that when the
migrtation flag drops and the fortress becomes open, we have the right
version. This should hopefully mitigate the need for multiple bounces
of the agent. Although, I don't mind as long as we're eventually
consistent. Charms should be able to handle that fact.